Update Literature.md
Added reference to arXiv:2009.03166v4 (NP B973 (2021) 115597).
The  range of  values of the couplings of the heavy Majorana neutrino N to Z as a function of the mass M,   for which it is possible to have successful leptogenesis, are 
given at the end of sections 3.3 and 3.4. There is comment about testing the considered leptogenesis  scenario at FCC-ee at the end of the Conclusions.